requirements.txt
Django==5.0.1
pillow==10.2.0
tesserocr==2.6.2
